refactor(work-items): drop review panel actions and orchestrator hook
Pre-commit hook ran. Total eslint: 0, total circular: 0
