+++ src/modules/ProjectManager/WorkItems/components/WorkItemDetail/index.tsx
@@ -21,7 +21,7 @@ import {
 } from "@src/types/core/workItem";
 
 import { getContextMenuItems } from "../../config";
-import { useWorkItemOrchestrator } from "../../hooks/useWorkItemOrchestrator";
+import { useWorkItemActiveSession } from "../../hooks/useWorkItemActiveSession";
 import { formatWorkItemShortId } from "../../workItemIdentity";
 import WorkItemContextMenu from "../WorkItemContextMenu";
 import { WorkItemDetailBody } from "./WorkItemDetailBody";
@@ -128,23 +128,9 @@ const WorkItemDetail: React.FC<WorkItemDetailProps> = ({
   const {
     activeAgentSessionId,
     activeAgentRole,
-    handleRetry,
-    handleCancelAgent,
-    handleAcceptAsIs,
-    handleCreateFollowUp,
     worktreePath,
     projectRepoPath,
-  } = useWorkItemOrchestrator({
-    workItem,
-    displayWorkItem,
-    repoPath,
-    projectSlug,
-    shortId,
-    onRefreshWorkItem,
-    onUpdateWorkItem,
-    hasPendingChanges,
-    handleSave,
-  });
+  } = useWorkItemActiveSession(workItem, repoPath);
 
   const { handleOpenFileDiff, handleOpenFileAtLine, handleReviewAllFiles } =
     useWorkItemFileActions(repoPath);
@@ -416,10 +402,6 @@ const WorkItemDetail: React.FC<WorkItemDetailProps> = ({
         onOpenSubItem={handleOpenSubItem}
         onUpdateWorkItem={handleLocalUpdate}
         onUpdateWorkItemImmediate={handleImmediateUpdate}
-        onCancelAgent={handleCancelAgent}
-        onRetry={handleRetry}
-        onAcceptAsIs={handleAcceptAsIs}
-        onCreateFollowUp={handleCreateFollowUp}
         onOpenSession={handleOpenSessionWithContext}
         onOpenFileDiff={handleOpenFileDiff}
         onOpenFileAtLine={handleOpenFileAtLine}
