--- src/modules/ProjectManager/WorkItems/hooks/useWorkItemOrchestrator/index.ts
@@ -1,2 +1,0 @@
-export { useWorkItemOrchestrator } from "./useWorkItemOrchestrator";
-export type { UseWorkItemOrchestratorOptions } from "./useWorkItemOrchestrator";
